% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{network_measure}
\alias{network_measure}
\title{Return network-level measures of a network.}
\usage{
network_measure(x)
}
\arguments{
\item{x}{a network}
}
\description{
\code{network_measure} calculate network-level measures of a network.
}
\details{
This function returns results of a network's network-level measures, including network size, density,
average degree, number of edges, reciprocity, transitivity, centralization, hierarchy,
number of components, connectedness
}
