% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{node_measure}
\alias{node_measure}
\title{Return node-level measures of a network.}
\usage{
node_measure(x)
}
\arguments{
\item{x}{a network}
}
\description{
\code{node_measure} calculate node-level measures of a network.
}
\details{
This function returns a network's node-level measures, namely every participant's in/outdegree, betweenness,
closeness.
}
