% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sna_plot}
\alias{sna_plot}
\title{Plot sociographs for networks.}
\usage{
sna_plot(x, y)
}
\arguments{
\item{x}{a network}

\item{y}{a matrix}
}
\description{
\code{sna_plot} plot sociograph for a network.
}
\details{
This function plots sociograph for a network in fruchtermanreingold mode, with node size and color representing
a participant's degree (indegree and outdegree) and edge width representing the frequency of interactions between
two participants
}
