% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week02_matrix}
\alias{week02_matrix}
\title{Week 02's participant to participant matrix format file}
\format{A square-format participant-by-participant matrix}
\source{
Fan Ouyang dissertation research, participants in week 02 discussion including 20 graduate students and 1 instructor
}
\usage{
week02_matrix
}
\description{
Week 02's participant to participant matrix format file
}
\examples{
\dontrun{
 week02_matrix
}
}
\keyword{datasets}
