% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week02_network}
\alias{week02_network}
\title{Week 02's participant to participant network format file}
\format{A network with multiple attributes, such as vertices, directed, bipartite, total edges, etc.}
\source{
Fan Ouyang dissertation research, participants in week 02 discussion including 20 graduate students and 1 instructor
}
\usage{
week02_network
}
\description{
Week 02's participant to participant network format file
}
\examples{
\dontrun{
 week02_network
}
}
\keyword{datasets}
