% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week08_matrix}
\alias{week08_matrix}
\title{Week 08's participant to participant matrix format file}
\format{A square-format participant-by-participant matrix}
\source{
Fan Ouyang dissertation research, participants in week 08 discussion including 20 graduate students and 1 instructor
}
\usage{
week08_matrix
}
\description{
Week 08's participant to participant matrix format file
}
\examples{
\dontrun{
 week08_matrix
}
}
\keyword{datasets}
