% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week08_network}
\alias{week08_network}
\title{Week 08's participant to participant network format file}
\format{A network with multiple attributes, such as vertices, directed, bipartite, total edges, etc.}
\source{
Fan Ouyang dissertation research, participants in week 08 discussion including 20 graduate students and 1 instructor
}
\usage{
week08_network
}
\description{
Week 08's participant to participant network format file
}
\examples{
\dontrun{
 week08_network
}
}
\keyword{datasets}
