% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week13_network}
\alias{week13_network}
\title{Week 13's participant to participant network format file}
\format{A network with multiple attributes, such as vertices, directed, bipartite, total edges, etc.}
\source{
Fan Ouyang dissertation research, participants in week 13 discussion including 20 graduate students and 1 instructor
}
\usage{
week13_network
}
\description{
Week 13's participant to participant network format file
}
\examples{
\dontrun{
 week13_network
}
}
\keyword{datasets}
