% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week14_matrix}
\alias{week14_matrix}
\title{Week 14's participant to participant matrix format file}
\format{A square-format participant-by-participant matrix}
\source{
Fan Ouyang dissertation research, participants in week 14 discussion including 20 graduate students and 1 instructor
}
\usage{
week14_matrix
}
\description{
Week 14's participant to participant matrix format file
}
\examples{
\dontrun{
 week14_matrix
}
}
\keyword{datasets}
