% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{week14_network}
\alias{week14_network}
\title{Week 14's participant to participant network format file}
\format{A network with multiple attributes, such as vertices, directed, bipartite, total edges, etc.}
\source{
Fan Ouyang dissertation research, participants in week 14 discussion including 20 graduate students and 1 instructor
}
\usage{
week14_network
}
\description{
Week 14's participant to participant network format file
}
\examples{
\dontrun{
 week14_network
}
}
\keyword{datasets}
